Moved find_overflights into prediction.c as find_passes
